+++ src/group/stabchain/mod.rs
@@ -3,7 +3,6 @@
 pub mod base;
 pub mod builder;
 pub mod element_testing;
-pub mod moved_point_selector;
 
 use crate::group::orbit::abstraction::TransversalResolver;
 use crate::group::Group;
@@ -11,7 +10,6 @@ use crate::perm::actions::SimpleApplication;
 use crate::perm::*;
 use base::Base;
 use builder::{Builder, BuilderStrategy};
-use moved_point_selector::MovedPointSelector;
 
 use crate::DetHashMap;
 
@@ -279,7 +277,6 @@ macro_rules! stabchain_tests {
     ($strategy:expr, $short:ident) => {
         mod $short {
             use crate::group::stabchain::builder::*;
-            use crate::group::stabchain::moved_point_selector;
             use crate::group::stabchain::{valid_stabchain, Stabchain};
             use crate::group::Group;
             use crate::perm::actions::*;
@@ -364,14 +361,14 @@ mod tests {
     stabchain_tests!(
         NaiveBuilderStrategy::new(
             SimpleApplication::default(),
-            moved_point_selector::LmpSelector::default()
+            crate::group::stabchain::base::selectors::LmpSelector::default()
         ),
         naive
     );
     stabchain_tests!(
         IFTBuilderStrategy::new(
             SimpleApplication::default(),
-            moved_point_selector::LmpSelector::default()
+            crate::group::stabchain::base::selectors::LmpSelector::default()
         ),
         ift
     );
@@ -381,7 +378,7 @@ mod tests {
             use rand::SeedableRng;
             RandomBuilderStrategy::new_with_params(
                 SimpleApplication::default(),
-                moved_point_selector::FmpSelector::default(),
+                crate::group::stabchain::base::selectors::FmpSelector::default(),
                 RandomAlgoParameters::default()
                     .rng(rand_xorshift::XorShiftRng::from_seed([43; 16])),
             )
